backend/pyproject.toml
[tool.poetry]
name = "classhub"
version = "1.0.1"
description = "ClassHub web application"
authors = ["Gabriel Rockson <thegabrielrockson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0.0"
djangorestframework = "^3.13.1"
django-extensions = "^3.1.5"
psycopg2-binary = "^2.9.3"
waitress = "^2.1.1"
whitenoise = {extras = ["brotli"], version = "^6.0.0"}
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
django-cors-headers = "^3.11.0"
django-filter = "^21.1"
Markdown = "^3.3.6"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
Werkzeug = "^2.1.1"
ptipython = "^1.0.1"
ipython = "^8.2.0"
bpython = "^0.22.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
